+++ cmd/hidden_lake/messenger/pkg/settings/settings.go
@@ -18,9 +18,8 @@ const (
 )
 
 const (
-	CDefaultInterfaceAddress     = "127.0.0.1:9591"
-	CDefaultIncomingAddress      = "127.0.0.1:9592"
-	CDefaultConnectionHLSAddress = "127.0.0.1:9592"
+	CDefaultInterfaceAddress = "127.0.0.1:9591"
+	CDefaultIncomingAddress  = "127.0.0.1:9592"
 )
 
 const (