+++ pkg/network/message/types.go
@@ -5,5 +5,5 @@ import "github.com/number571/go-peer/pkg/payload"
 type IMessage interface {
 	GetHash() []byte
 	GetPayload() payload.IPayload
-	GetBytes() []byte
+	ToBytes() []byte
 }