+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - Update `hidden_lake/service`: messageSize (4 << 20) -> (4 << 10)
 - Update `Makefiles`: append .exe extenstion to windows compile
+- Update `hidden_lake/messenger`: edit CDefaultConnectionHLSAddress -> hls_settings.CDefaultHTTPAddress
 
 <!-- ... -->